added failing test for proof of concept
The value of foreign key is not passed down to user_id during approval as it should
