+++ tests/Feature/Scopes/ApprovalStateScopeTest.php
@@ -6,8 +6,8 @@ use Cjmellor\Approval\Events\ModelRejected;
 use Cjmellor\Approval\Events\ModelSetPending;
 use Cjmellor\Approval\Models\Approval;
 use Cjmellor\Approval\Tests\Models\FakeModel;
+use Cjmellor\Approval\Tests\Models\FakeUser;
 use Illuminate\Support\Facades\Event;
-use Illuminate\Support\Facades\Schema;
 
 test('Check if an Approval Model is approved', closure: function (): void {
     $this->approvalData = [
@@ -202,27 +202,7 @@ test(description: 'A Model can be Postponed unless a condition is met', closure:
 });
 
 test(description: 'The model approver is listed correctly', closure: function () {
-    Schema::create('fake_users', callback: function (Illuminate\Database\Schema\Blueprint $table) {
-        $table->id();
-        $table->string(column: 'name');
-        $table->string(column: 'email')->unique();
-        $table->string('password');
-    });
-
-    class FakeUser extends \Illuminate\Foundation\Auth\User
-    {
-        protected $guarded = [];
-
-        protected $table = 'fake_users';
-
-        public $timestamps = false;
-    }
-
-    $user = FakeUser::create([
-        'name' => 'Chris Mellor',
-        'email' => 'chris@mellor.pizza',
-        'password' => 'password',
-    ]);
+    $user = FakeUser::create($this->fakeUserData);
 
     $this->be($user);
 
@@ -232,6 +212,27 @@ test(description: 'The model approver is listed correctly', closure: function ()
     $approval->approve();
 
     expect($approval)->fresh()->audited_by->toBe(expected: $user->id);
+});
 
-    Schema::dropIfExists('fake_users');
+test(description: 'The model foreign key is set correctly', closure: function () {
+    $user = FakeUser::create($this->fakeUserData);
+
+    $this->be($user);
+
+    $fakeModelData = [
+        ...$this->fakeModelData,
+        'user_id' => $user->id
+    ];
+
+    FakeModel::create($fakeModelData);
+
+    $approval = Approval::first();
+    $approval->approve();
+
+    $this->assertDatabaseHas(
+        'fake_models',
+        [
+            'user_id' => $user->id
+        ]
+    );
 });
